I've made the following changes to AnimationWidget:
1) Have it check the state of the AnimationControl's
looping and direction.  To do the direction, I added
getDirection() to AnimationControl, AnimationControlJ2D,
and AnimationControlJ3D.  There was a setDirection, but
no getDirection.

2) I changed the slider so its base is 1 instead of 0.
This goes along with the recent changes Bill made to the
DisplayRenderer so the values listed now say 1 of x, 2 of x,
etc instead of 0 of x, 1 of x.

3) I had to set the maximum size of the Stop/Go button to
be the max size of the "Step" button because if it came up
as Go, Stop would show as ....  This was just rearranging
some code that Bill already had there.

4) Added some javadoc.