Undo Bill's commit (which essentially reversed my commit and broke SelectRangeWidget for collaboration)
The *real* problem is the race condition between the main thread,
which is busy initializing the RangeControl, and the Action thread,
which is in the midst of a DisplayImpl.doAction().  Occasionally,
RangeControl.getRange() gets called before the control is initialized
and the default range of [0.0,0.0] is returned (and thus no points
are displayed.)

Bill's commit avoids the race condition by scheduling another doAction().

This commit avoids the race by constructing RangeControl with a range
of [NaN,NaN], thus ensuring that all data are displayed.