There's no reason this class needs to be 'abstract', so remove that qualifier.
Rename DEFAULT_TABLE_LENGTH to DEFAULT_NUMBER_OF_COLORS and make it 'public'

In syncControl(), I was just using the reference to the other control's table.
This works in C/S mode (because the table has been serialized and is this
a copy of the original control's table), but falls down as soon as a local
control is synced (because then you've got two controls referencing the same
table.)  Therefore, I need to always copy the table.