Make the whole fleet of colormap-related widgets use a ColorControl as their source.
I *REALLY* wanted to avoid this, but the way these controls were written
meant that 2 controlChanged() events (and a couple of table rewrites)
were done for each change (and every mouse drag contains at least two
changes) and that makes event loops a near certainty in the collaborative
world.  I have some change damping code for precisely this reason, but
*that* causes problems for binary-values Controls. so the damping code
has to go, which means these widgets have to be fixed.