Test application for binary file reader/writer
Without any arguments, it builds one of every type of data object,
writes it to a binary file, reads the binary file in and compares the
resulting Data object with the original.

Any arguments are treated as files, which are read with
visad.data.DefaultFamily, written as binary files, read in and
compared with the objects created from the original files.