progress
Added new methods to AxisScale:

  - set/getMajorTickSpacing  (to change the major tick spacing)
  - set/getMinorTickSpacing  (to set minor tick spacing)
  - set/getFont              (to set the label font)
  - setSnapToBox             (to set the label base line against the box)
  - setAutoComputeTicks      (to automatically calculate ticks, size, etc)

Size of labels is based on font size (default with no font is same as
before).  Added new setScale signature to DisplayRenderer and subclasses,
added javadoc to makeAppearance method of ShadowTypeJ3D.

Still to go:

display labels on right side of box
set the labels that are displayed
others?

Let me know if there are problems.