Initial commit.
This class supports a VisAD Set that comprises double-precision data on
a 1-D manifold with no topology (i.e. nothing should be inferred from
the order of the data).  This class is the double-precision analogue of
class List1DSet.

I had planned to create the interface IrregularDoubleSet and the class
Irregular1DDoubleSet, but this class suffices for my needs (importing double-
precision, netCDF attributes).