AreaDirectory - return "original" source type if data source type is "VISR" and "original" source type in AREA directory is non-zero.
AreaFile - Added AREA directory word AD_SRCTYPEORIG as the "original"
source type.  When the "String" constructor is called, if it begins
with "adde://" and ends with "image?", the invoke the GetAreaGUI to
allow the user to build a real URL; this will work with any VisAD (or
other) application.

GetAreaFile - command line application to get AREA files from an ADDE
server and write them on local disk; all parameters may be on the
command line, or using the -gui option, invoke the GetAreaGUI.

GetAreaGUI - a GUI for exploring an ADDE server for image data. Will
provide a registered ActionListener with the complete URL, assembled
from the user's choices.