Added drawString() method to put HersheyFont text anywhere in the Display's domain.  Need a rotate= keyword in addition to the start/base/up keywords.
Fixed color= keyword everywhere so it can either be a java.awt.Color
or just the name of one of the AWT colors.

Created internal RealTypes (py_shape_type and py_text_type) and use
these to make ScalarMaps which are added to every Display when they
are created.

Need more commentary.