Added HersheyFonts to TextControl.  Changes made to ShadowType to support this.  example/Text69 can now accept a 2nd parameter - the name of a HersheyFont to use; otherwise, it still uses serif.  Try "futural", "timesrb" or "cursive"...or the others.   Ex:    java Test69 futural sphere
TextControl.setFont() changed to take an Object, which is then checked
to see if it's a java.awt.Font or a visad.util.HersheyFont.  New method
getHersheyFont() added to TextControl.  If the font has been set to a
HersheyFont, this will return it; otherwise returns null (same rule as
for getFont()).