Changed the internal hashtable into a memory-sensitive, canonicalizing cache that will not prevent a ScalarType from being garbaged-collected when it is no longer strongly- or softly-referenced outside this class.  The cache is checked for garbaged-collected instances (which are subsequently removed from the cache) upon invocation of any relevant method.
Modified the readResolve() method: if a deserialized instance is
incompatible with a previously-existing instance with the same name,
then an exception is now thrown rather than the deserialized instance
returned.  Also, the method is now public so that it is always invoked
and final so that it can't be evaded.

Tightened-up synchronization around the hashtable and alias translation
tables.

The modification passed all the ScalarType tests I threw at them, but
they should probably be tested some more.