Add support for InputEvents tied to DisplayEvents.  Whenever a MOUSE type DisplayEvent occurs, you can call getInputEvent() to get the MouseEvent describing the details of the event.  A getModifiers() method has also been added (it's similar to getX() and getY()).
KeyEvents are not yet supported but could easily be added if
KeyboardBehaviors were changed to start triggering DisplayEvents when
keys are pressed.

Also, it should be easy to change MouseHelper to trigger a DisplayEvent
on MOUSE_DRAGGED events (I think people were asking for something like
this a while ago).