Lose BoxLayout in favor of BorderLayout
BoxLayout will generally lay thing out from top to bottom (or left to right)
but occasionally will lay things out in the reverse direction.  When it
does things in reverse, it can get a completely different layout, and thus
cause the GUI to shift radically.  The biggest problem is when this radical
shift throws GUI elements completely off the bottom of the window.