Restructure to make this subclass JPanel instead of JDialog.  Use an internal JDialog, unless told otherwise by the new Constructor's last parameter.  Remove constructor that claims the parent was a Dialog, since there is really no need for that one....and it wouldn't work right anyway.
This was done so that applications like the IDV can simply add this
entire interface as a JPanel within their own world.

Also, move the properties file from the 'current directory' to the
System.getProperties("user.home") directory, so it can be used no
matter where the user starts their program.