Fix a nasty and early bug in VisAD. The problem is that the full Linear2DSet constructor allows applications to specify Units for component Linear1DSets in two ways: 1. in a Unit[] argument and 2. in the Units of the component Linear1DSet[] argument.
This changes Linear2DSet, Linear3DSet and LinearNDSet
(and this changes the behavior of LinearLatLonSet which
extends Linear2DSet) to ensure that the Units of the
component Linear1DSets are consistent with the Units
returned by getSetUnits() (and stored way up the class
hierarchy in Set.SetUnits). This also entails some
changes to IntegerSet classes.

The policy is to determine Units according to this priority:
1. Unit[] argument passed to constructor, if non-null.
2. Units in component Linear1DSet[] argument, for constructor
signatures that have have that argument.
3. Default Units in SetType.

Such forced consistency may break some applications that
rely on the current anomolous behavior. Note especially
that if an application calls a Linear2DSet, Linear3DSet,
LinearNDSet or LinearLatLonSet constructor with
'Linear1DSet[] sets' and 'Unit[] units' arguments where
units[i] is not equal to the Unit of sets[i], then the
actual values of the Linear1DSets will be converted from
the Unit of sets[i] to units[i].