Add MetadataReader interface for file forms that wish to provide additional metadata from the file. Uses a hashtable with each piece of metadata as a named key.
Adapted BioRadForm to use MetadataReader.

Removed BioRadCachingForm, because it was not being used, and it was
broken (testing with the latest version before these changes, it throws
ClassCastException). Since I am unsure how to adapt it to work with the
cleaned up version of BioRadForm, I am removing it. It is doubtful that
anyone else is using it, or else they would have complained about its
currently broken state.