Use GIFForm before LociForm. The rationale is as follows:
Many GIFs, such as those in visad/ss, use 8-bit indexed color.

LociForm uses ImageIO to read the pixel values correctly, returning type
((ImageElement, ImageLine) -> value0). But the associated indexed color
table is unfortunately ignored in the conversion from BufferedImage to
VisAD object.

GIFForm, on the other hand, uses an old-school Toolkit.getImage call,
which retrieves a non-buffered java.awt.Image object, which is then
translated into a VisAD object with DataUtility.makeImage (which assumes
the image is RGB color).

So, GIFForm results in a larger data object, but which displays correctly
in the VisAD SpreadSheet with default mappings to Red, Green and Blue.
For compatibility and consistency, we will use GIFForm.