Re-order Data transform path for isosurface rendering *only* when spatial_set cannot be formed as a result of a display coord_sys transform, eg. a global model mapped to a GVAR navigated display.
This involves adding a new signature, makeIsoSurfaceMissingSpatial,
which is invoked by makeContour for the case above.  The new
method is essentially the same as 'makeIsoSurface' except that the
isosurface is computed in data coordinates and then map scaling,
coordinate transforms and spatial offsets are applied to  the
geometry rather than the sampling domain.

This new method is necessary for two reasons:
(1) testing revealed that the marching-cube algorithm will not
    work consistently if there are missing samples.  I wish it
    would have.

(2) The scaling and transforms of the resulting geometry must
    be applied before the normals are computed for consistency.

I've tested this pretty thoroughly in the IDV and the resulting
isosurfaces look consistent with those generated via the normal
path.