Tom W. discovered that the jar target for making the distribution jar file did not work under windows, so I had to change the incredibly cumbersome manner in which the dist jar file was created.
There is now a patternset at the top where the file patterns for all the
files to be included in the jar are defined. This may need to be fine
tuned some because now it's pretty general, including just about all class
files.

Also, I should note that the changes were tested on my MacBook and a
Parallels VM running Windows XP both running Ant 1.6.2 on top of Java 1.5
and everything worked fine.