The jar.src target now contains the same files as produced by the make file except that it contains this build.xml file.
The patternsets for the individual packages are now the only source for
determining the java related source for the jar, jar.src, and javadoc
targets. The patternsets specify all files when added to a fileset in a
targets and from there they are filtered for specific needs.
NOTE: There may be a better way to do this, but I haven't figured it out
yet.

Added patternsets for including non-java files in the source and
distribution jars.