Simpler copyright update script.
It does not do as much as copyright.pl does, but I believe it is
sufficient for the time being. It also makes minimal assumptions about
the structure of the copyright blurbs. However, it is more dangerous
than copyright.pl because it does not report the status of source files:
which copyright they have, whether they have one at all, etc.

This script is what was used to update the copyright to 2011.