Fix problem with getting "localhost".
On my Linux machine, the hostname will stay the same regardless of the
network I'm on, while macOS will change the hostname (particularly
evident when using wifi).

This leads to attempting to resolve a potentially invalid hostname in an
apparent effort to get the current IP (note: which is not the same thing
as loopback/localhost). This commit simply grabs the local address from
an extant Socket.